+++ tests/rest/client/test_sliding_sync.py
@@ -4,11 +4,12 @@ from synapse.server import HomeServer
 from synapse.types import JsonDict
 from synapse.util import Clock
 
-from tests.rest.client.test_sendtodevice import SendToDeviceTestCase
+from tests.rest.client.test_sendtodevice_base import SendToDeviceTestCaseBase
+from tests.unittest import HomeserverTestCase
 
 
 # Test To-Device messages working correctly with the `/sync/e2ee` endpoint (`to_device`)
-class SlidingSyncE2eeSendToDeviceTestCase(SendToDeviceTestCase):
+class SlidingSyncE2eeSendToDeviceTestCase(SendToDeviceTestCaseBase, HomeserverTestCase):
     def default_config(self) -> JsonDict:
         config = super().default_config()
         # Enable sliding sync
@@ -19,4 +20,4 @@ class SlidingSyncE2eeSendToDeviceTestCase(SendToDeviceTestCase):
         # Use the Sliding Sync `/sync/e2ee` endpoint
         self.sync_endpoint = "/_matrix/client/unstable/org.matrix.msc3575/sync/e2ee"
 
-    # See SendToDeviceTestCase for tests
+    # See SendToDeviceTestCaseBase for tests
